requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.4.0
backcall==0.2.0
backoff==2.2.1
bidict==0.22.1
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
cffi==1.15.1
chainlit==0.6.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.6
comm==0.1.4
contourpy==1.1.0
crashtest==0.4.1
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.13
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.7
dulwich==0.21.5
executing==1.2.0
fastapi==0.97.0
fastapi-socketio==0.0.10
filelock==3.12.2
filetype==1.2.0
fonttools==4.42.1
frozenlist==1.4.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.1
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.25.0
ipython==8.14.0
jaraco.classes==3.3.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keyring==23.13.1
kiwisolver==1.4.4
langchain==0.0.246
langsmith==0.0.15
lockfile==0.12.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
more-itertools==10.0.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.6
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
packaging==23.1
pandas==2.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipe==2.2
pkginfo==1.9.6
platformdirs==3.9.1
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
prisma==0.9.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-graphql-client==0.4.3
python-socketio==5.8.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.0
rapidfuzz==2.15.1
referencing==0.30.0
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
subgrounds==1.8.1
syncer==2.0.3
tenacity==8.2.2
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trove-classifiers==2023.7.6
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2024.1
uptrace==1.18.0
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.24.2
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
wrapt==1.15.0
xattr==0.10.1
yarl==1.9.2
zipp==3.16.2
